# tammy-ai-server-main/docs/swagger.yaml
definitions:
...
  github_com_yeongin-ji_tammy-ai-server_internal_dto.ChatRequest:
    properties:
      history:
        description: Prior turns in chronological order, oldest first. May be empty.
        items:
          $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ChatTurn'
        type: array
      nickname:
        description: Optional display name so Tammy can address the user naturally.
        example: 우당탕탕
        type: string
      userId:
        example: 12
        type: integer
      userMessage:
        example: 오늘 회사에서 일이 너무 많아서 힘들고 지쳤어
        type: string
    required:
    - userId
    - userMessage
    type: object
  github_com_yeongin-ji_tammy-ai-server_internal_dto.ChatResponse:
    properties:
      emotion:
        $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.EmotionStatus'
      replyText:
        example: 오늘 정말 많이 힘들었구나. 괜찮아, 우리 잠깐 바람 쐬러 갈까?
        type: string
    type: object
  github_com_yeongin-ji_tammy-ai-server_internal_dto.ChatTurn:
    properties:
      createdAt:
        description: RFC3339 timestamp. Optional; used only to give the model a sense
          of pacing.
        example: "2026-08-05T21:14:00Z"
        type: string
      role:
        description: Either "user" or "tammy".
        enum:
        - user
        - tammy
        example: user
        type: string
      text:
        example: 오늘 야근해서 너무 지쳤어
        type: string
    required:
    - role
    - text
    type: object
...
  github_com_yeongin-ji_tammy-ai-server_internal_dto.EmotionStatus:
    properties:
      motionType:
        description: Tammy's reaction animation, one of the MotionType values.
        example: PAT_PAT_HEAD
        type: string
      state:
        description: |-
          One of HAPPY, SAD, ANGRY, STRESSED, CALM — matching the service server's
          EmotionState enum so it can be written straight into emotion_logs.
        example: STRESSED
        type: string
    type: object
  github_com_yeongin-ji_tammy-ai-server_internal_dto.ErrorResponse:
    properties:
      code:
        example: AI_MODEL_ERROR
        type: string
      message:
        example: AI 모델 호출에 실패했습니다.
        type: string
    type: object
...
  github_com_yeongin-ji_tammy-ai-server_internal_dto.NutritionItem:
    properties:
      caloriesKcal:
        example: 243
        type: number
      carbohydrateG:
        example: 12.4
        type: number
      confidence:
        description: Model confidence in the range [0, 1].
        example: 0.81
        type: number
      fatG:
        example: 13.8
        type: number
      mineralPercent:
        example: 42
        type: integer
      name:
        example: 김치찌개
        type: string
      proteinG:
        example: 15.2
        type: number
      servingSizeG:
        description: Serving the figures below describe, in grams.
        example: 400
        type: number
      sources:
        description: |-
          Where the figures came from. May be empty when the model fell back to
          general knowledge, in which case Confidence will be low.
        items:
          $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.NutritionSource'
        type: array
      vitaminPercent:
        example: 35
        type: integer
    type: object
  github_com_yeongin-ji_tammy-ai-server_internal_dto.NutritionLookupRequest:
    properties:
      foodNames:
        description: Food names to research. Must contain at least one entry.
        example:
        - 김치찌개
        - 공기밥
        items:
          type: string
        minItems: 1
        type: array
    required:
    - foodNames
    type: object
  github_com_yeongin-ji_tammy-ai-server_internal_dto.NutritionLookupResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.NutritionItem'
        type: array
    type: object
  github_com_yeongin-ji_tammy-ai-server_internal_dto.NutritionSource:
    properties:
      publisher:
        example: 식품의약품안전처
        type: string
      title:
        example: 식품영양성분 데이터베이스 - 김치찌개
        type: string
      url:
        example: https://various.foodsafetykorea.go.kr/nutrient/
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    웰니스 펫 에이전트 '타미'의 AI 서버입니다. 서비스 서버와 내부 통신하며, 상태를 저장하지 않습니다.
    Cloud Run은 인증되지 않은 호출을 허용하며, 접근 제어는 `X-Internal-Api-Key` 헤더가 담당합니다.
  license:
    name: MIT
  title: TAMMY AI Server API
  version: "1.0"
paths:
...
  /v1/chat/process:
    post:
      consumes:
      - application/json
      description: |-
        현재 발화와 이전 대화 히스토리를 받아 타미 페르소나로 답변하고, **사용자 발화에 대한 감정 분석 결과**를 함께 반환합니다.
        이 서버는 상태를 저장하지 않으므로 대화 맥락이 필요하면 `history`에 이전 턴을 담아 보내야 합니다.
        `history`는 오래된 순서대로 정렬되어야 하며, 최근 30턴만 모델에 전달됩니다.
        `emotion.state`는 서비스 서버의 `EmotionState` enum(HAPPY, SAD, ANGRY, STRESSED, CALM)과 동일합니다.
        `emotion.motionType`은 타미의 반응 모션입니다.
      parameters:
      - description: 현재 발화와 대화 히스토리
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ChatResponse'
        "400":
          description: 요청 형식 오류
          schema:
            $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ErrorResponse'
        "502":
          description: 모델 오류
          schema:
            $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ErrorResponse'
        "504":
          description: 모델 응답 지연
          schema:
            $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ErrorResponse'
      summary: 타미 대화 처리 및 감정 분석
      tags:
      - chat
  /v1/nutrition/lookup:
    post:
      consumes:
      - application/json
      description: |-
        음식 이름 배열을 받아 웹 검색을 통해 각 음식의 칼로리, 탄수화물, 단백질, 지방, 비타민/무기질 충족률을 조사하고 출처를 함께 반환합니다.
        응답의 `items`는 요청한 `foodNames`와 **같은 순서, 같은 개수**로 반환되므로 인덱스로 짝지을 수 있습니다.
        검색으로 확인하지 못한 음식은 수치가 0, `confidence`가 0으로 채워집니다.
        웹 검색과 구조화를 위해 모델을 두 번 호출하므로 다른 엔드포인트보다 응답이 느립니다.
      parameters:
      - description: 조회할 음식명 목록
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.NutritionLookupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.NutritionLookupResponse'
        "400":
          description: 요청 형식 오류
          schema:
            $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ErrorResponse'
        "502":
          description: 모델 오류
          schema:
            $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ErrorResponse'
        "504":
          description: 모델 응답 지연
          schema:
            $ref: '#/definitions/github_com_yeongin-ji_tammy-ai-server_internal_dto.ErrorResponse'
      summary: 음식명 배열로 영양 정보 조회
      tags:
      - nutrition
...
securityDefinitions:
  InternalApiKey:
    description: 서비스 서버와 공유하는 내부 API 키
    in: header
    name: X-Internal-Api-Key
    type: apiKey
swagger: "2.0"